Reduce duplication in macro/rule
